# add performance optimizations for 60-80% faster audio generation