--- a/.gitignore
+++ b/.gitignore
@@ -175,3 +175,6 @@ tags
 /checkpoints/
 exp
 .gradio/
+
+# Clacky-related configuration
+.clackyai/